requirements.txt
fastapi
uvicorn
python-multipart
pyjwt
passlib
bcrypt
